Fix manual data cleaning when new pattern has a different order of using entities
The updated pattern might have a different list of entities and may have a different list of entities,
or the same list of entities with different order. So when we try to appy the changes to other examples
that share the same old pattern, we need to first update the entity list to match the new pattern.